fix(cri): Fix unit testing errcheck and remove unused mock file
Some files has been renamed along the way
